Add files via upload
Multiple Joy! is a multiple choice game for kids. It retrieves a list of names and words from a CSV file that contains these names/words and the categories to which they belong. From a combobox you can select the game gategory (e.g. Disney characters, furniture, vehicles, etc.). The program creates two sets of radio buttons one of which shows the names/words, and a second set which shows the images of the characters/things named in the first set. When a name/word matches its picture, a larger version of this picture is shown on a canvas.