Add files via upload
A little race game with no GUI. The user has to enter values for fuel, speed, and laps. Then chooses the vehicles (car, motorbike or truck) to compete. The race results are automatically generated. The output prints on the terminal the lap stats for each racer, and the final results. If any vehicle runs out of gas before reaching the finish line, it is disqualified. Fuel and speed are variable depending on the vehicle type.