Add files via upload
This is a text generator. It was named after my daughter. I made it for her. As long as one presses any key on the keyboard, the program inserts random words into a text box, as if the child was actually writing a story. Additionally, it inserts (also randomly) punctiation marks so as to resemble an actual story. The result is really funny! You can even save the text into a TXT document with a randomly generated title. I took the words from both English and Spanish versions of Alice's Adventures in Wonderland.